fix: ignoring snapshots with pending writes
When writing a firestore.FieldValue.serverTimestamp() value to a doc, it
takes two snapshots until that value is properly received by the
snapshot listener. This delay causes null pointer exceptions to be
thrown.
In order to avoid them, the snapshot contains info (metadata) indicating
whether writes are pending or not. This is the chosen strategy to avoid
this null pointers.