fix: hardcoded bg color to match cosmic theme
The web interface doesn't have any scroll, so the mismatch between the
theme and the UI's background color is not visible. But when installing
the app on a phone as a PWA, there is over-scroll, which shows this
mismatch. If the theme is changed, then the mismatch is visible as well,
but with this approach we at least fix it somewhere.